refactor/user_repository_test.goの内容をモジュール化しリファクタリング
